git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@192666 4525493e-7705-40b1-a816-d608a930855b
Former-commit-id: 414c0e186506c4cec7f1c3068afeaf3b563474ee

+++ Online/Online/DefHLTUtils/src/FileDeleter.cpp
@@ -92,6 +92,7 @@ class CommandInfo : public DimUpdatedInfo
         std::string shcomm;
         for (i=0;i<splt2->size();i++)
         {
+          if (splt2->at(i).size() == 0) continue;
           shcomm = "sudo rm -rf /localdisk/hlt1/Run_*"+splt2->at(i)+"_*.mdf";
           printf("executing command %s\n",shcomm.c_str());
           int status = ::system(shcomm.c_str());
